読みだすデータを追加

+++ src/Ground/Teleplot/FlightModule/FlightModule.cpp
@@ -8,11 +8,11 @@
 
 StaticJsonDocument<4096> packet;
 
-uint16_t separation1ProtectionTime = 8605;
-uint16_t separation1ForceTime = 11605;
-uint16_t separation2ProtectionTime = 1000;
-uint16_t separation2ForceTime = 1000;
-uint16_t landingTime = 30305;
+uint32_t separation1ProtectionTime = 8605;
+uint32_t separation1ForceTime = 11605;
+uint32_t separation2ProtectionTime = 1000;
+uint32_t separation2ForceTime = 1000;
+uint32_t landingTime = 30305;
 
 void setup()
 {
